refactor(month-body): split month-body into multiple files
